+++ Modules/DateResolver.psm1
@@ -177,7 +177,7 @@ function Get-DateMatchResult {
 
     $date = & $Parser $match
     if ($null -eq $date) {
-        return $null
+        return New-CaptureDateResult -Success $false -DateTime $null -Source 'None' -Pattern $PatternName -Warnings @("Filename matched pattern '$PatternName' but contains an invalid date.")
     }
 
     return New-CaptureDateResult -Success $true -DateTime $date -Source 'FileName' -Pattern $PatternName -Warnings @()
@@ -190,11 +190,49 @@ function New-DateTimeFromParts {
         [string]$Day,
         [string]$Hour,
         [string]$Minute,
-        [string]$Second
+        [string]$Second,
+        [string]$Millisecond = '0'
     )
 
+    $currentYear = (Get-Date).Year
+    $yearValue = [int]$Year
+    $monthValue = [int]$Month
+    $dayValue = [int]$Day
+    $hourValue = [int]$Hour
+    $minuteValue = [int]$Minute
+    $secondValue = [int]$Second
+    $millisecondValue = [int]$Millisecond
+
+    if ($yearValue -lt 2000 -or $yearValue -gt ($currentYear + 1)) {
+        return $null
+    }
+
+    if ($monthValue -lt 1 -or $monthValue -gt 12) {
+        return $null
+    }
+
+    if ($dayValue -lt 1 -or $dayValue -gt 31) {
+        return $null
+    }
+
+    if ($hourValue -lt 0 -or $hourValue -gt 23) {
+        return $null
+    }
+
+    if ($minuteValue -lt 0 -or $minuteValue -gt 59) {
+        return $null
+    }
+
+    if ($secondValue -lt 0 -or $secondValue -gt 59) {
+        return $null
+    }
+
+    if ($millisecondValue -lt 0 -or $millisecondValue -gt 999) {
+        return $null
+    }
+
     try {
-        return [datetime]::new([int]$Year, [int]$Month, [int]$Day, [int]$Hour, [int]$Minute, [int]$Second)
+        return [datetime]::new($yearValue, $monthValue, $dayValue, $hourValue, $minuteValue, $secondValue, $millisecondValue)
     } catch {
         return $null
     }
@@ -212,7 +250,10 @@ function Get-VideoDateFromFileName {
     $baseName = [System.IO.Path]::GetFileNameWithoutExtension($Path)
 
     $patterns = @(
-        @{ Name = 'AndroidPrefix'; Regex = '^(?:VID|IMG|PXL)_(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})_(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})(?:\d{3})?$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
+        @{ Name = 'Imou_yyyyMMddHHmmssfff_prefix'; Regex = '^(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})(?<fff>\d{3})_'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value $m.Groups['fff'].Value } }
+        @{ Name = 'Insta360_VID_yyyyMMdd_HHmmss_suffix'; Regex = '^VID_(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})_(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})(?:_.*)+$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
+        @{ Name = 'VID_yyyyMMdd_HHmmss'; Regex = '^VID_(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})_(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
+        @{ Name = 'AndroidPrefix'; Regex = '^(?:IMG|PXL)_(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})_(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})(?:\d{3})?$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
         @{ Name = 'BasicCompact'; Regex = '^(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})[_-](?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
         @{ Name = 'BasicIsoCompact'; Regex = '^(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})T(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})Z?$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
         @{ Name = 'DashedSpaced'; Regex = '^(?<y>\d{4})[-_.](?<m>\d{2})[-_.](?<d>\d{2})[ _](?<hh>\d{2})[.\-_:](?<mm>\d{2})[.\-_:](?<ss>\d{2})$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
@@ -237,14 +278,40 @@ function Get-VideoDateFromFileName {
         @{ Name = 'GoPro'; Regex = '^(?:GOPR\d{4}|GX\d{6})_(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})_(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})$'; Parser = { param($m) New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value } }
     )
 
+    $warnings = New-Object System.Collections.Generic.List[string]
     foreach ($pattern in $patterns) {
         $result = Get-DateMatchResult -BaseName $baseName -PatternName $pattern.Name -RegexPattern $pattern.Regex -Parser $pattern.Parser
         if ($null -ne $result) {
-            return $result
+            if ($result.Success) {
+                return $result
+            }
+
+            foreach ($warning in @($result.Warnings)) {
+                $warnings.Add($warning)
+            }
         }
     }
 
-    return New-CaptureDateResult -Success $false -DateTime $null -Source 'None' -Pattern $null -Warnings @('Filename does not match any supported date pattern.')
+    $generic17Digit = Get-DateMatchResult -BaseName $baseName -PatternName 'Generic_yyyyMMddHHmmssfff' -RegexPattern '(?<y>\d{4})(?<m>\d{2})(?<d>\d{2})(?<hh>\d{2})(?<mm>\d{2})(?<ss>\d{2})(?<fff>\d{3})' -Parser {
+        param($m)
+        New-DateTimeFromParts $m.Groups['y'].Value $m.Groups['m'].Value $m.Groups['d'].Value $m.Groups['hh'].Value $m.Groups['mm'].Value $m.Groups['ss'].Value $m.Groups['fff'].Value
+    }
+
+    if ($null -ne $generic17Digit) {
+        if ($generic17Digit.Success) {
+            return $generic17Digit
+        }
+
+        foreach ($warning in @($generic17Digit.Warnings)) {
+            $warnings.Add($warning)
+        }
+    }
+
+    if ($warnings.Count -eq 0) {
+        $warnings.Add('Filename does not match any supported date pattern.')
+    }
+
+    return New-CaptureDateResult -Success $false -DateTime $null -Source 'None' -Pattern $null -Warnings @($warnings)
 }
 
 function Resolve-VideoCaptureDate {